+++ app/core/interfaces/color/lib/color.js
